+++ packages/conan/recipes/ninja/conanfile.py
@@ -1,20 +1,16 @@
 # Copyright (c) Contributors to the conan-center-index Project. All rights reserved.
 # Copyright (c) Contributors to the aswf-docker Project. All rights reserved.
 # SPDX-License-Identifier: MIT
+#
+# From: https://github.com/conan-io/conan-center-index/blob/5aa8fda35e023ceeaefb5a670ae9c0c3ca4fdf25/recipes/ninja/all/conanfile.py
 
-from conan import ConanFile
-from conan.tools.cmake import CMake, CMakeDeps, CMakeToolchain, cmake_layout
-from conan.tools.files import (
-    apply_conandata_patches,
-    collect_libs,
-    copy,
-    export_conandata_patches,
-    get,
-    rmdir,
-)
-from conan.tools.microsoft import is_msvc
+from conan import ConanFile, conan_version
+from conan.tools.cmake import CMake, CMakeToolchain, cmake_layout
+from conan.tools.files import copy, get, replace_in_file
 from conan.tools.scm import Version
 import os
+
+required_conan_version = ">=1.52.0"
 
 
 class NinjaConan(ConanFile):
@@ -22,15 +18,10 @@ class NinjaConan(ConanFile):
     package_type = "application"
     description = "Ninja is a small build system with a focus on speed"
     license = "Apache-2.0"
-    url = "https://github.com/AcademySoftwareFoundation/aswf-docker"
+    url = "https://github.com/conan-io/conan-center-index"
     homepage = "https://github.com/ninja-build/ninja"
     topics = ("ninja", "build")
-    settings = (
-        "os",
-        "arch",
-        "compiler",
-        "build_type",
-    )
+    settings = "os", "arch", "compiler", "build_type"
 
     def layout(self):
         cmake_layout(self, src_folder="src")
@@ -42,15 +33,16 @@ class NinjaConan(ConanFile):
         get(self, **self.conan_data["sources"][self.version], strip_root=True)
         cmakelists = os.path.join(self.source_folder, "CMakeLists.txt")
 
+        #prevent re2c (which is optional and not needed) from being used
+        replace_in_file(self, cmakelists, "if(RE2C)", "if(FALSE)")
+
     def generate(self):
         tc = CMakeToolchain(self)
-        tc.variables["BUILD_TESTING"] = "OFF"
+        tc.variables["BUILD_TESTING"] = False
         if self.settings.os == "Linux" and "libstdc++" in self.settings.compiler.libcxx:
             # Link C++ library statically on Linux so that it can run on systems
             # with an older C++ runtime
-            tc.cache_variables[
-                "CMAKE_EXE_LINKER_FLAGS"
-            ] = "-static-libstdc++ -static-libgcc"
+            tc.cache_variables["CMAKE_EXE_LINKER_FLAGS"] = "-static-libstdc++ -static-libgcc"
         tc.generate()
 
     def build(self):
@@ -59,17 +51,15 @@ class NinjaConan(ConanFile):
         cmake.build()
 
     def package(self):
-        copy(
-            self,
-            "COPYING",
-            src=self.source_folder,
-            dst=os.path.join(self.package_folder, "licenses", self.name),
-        )
+        # ASWF: license files in per package folder
+        copy(self, "COPYING", src=self.source_folder, dst=os.path.join(self.package_folder, "licenses", self.name))
         cmake = CMake(self)
         cmake.install()
 
     def package_info(self):
         self.cpp_info.includedirs = []
         self.cpp_info.libdirs = []
-        self.env_info.PATH.append(os.path.join(self.package_folder, "bin"))
-        self.env_info.CONAN_CMAKE_GENERATOR = "Ninja"
+        # TODO: to remove in conan v2
+        if Version(conan_version).major < 2:
+            self.env_info.PATH.append(os.path.join(self.package_folder, "bin"))
+            self.env_info.CONAN_CMAKE_GENERATOR = "Ninja"