+++ packages/conan/recipes/nlohmann_json/conanfile.py
@@ -1,0 +1,60 @@
+# Copyright (c) Contributors to the conan-center-index Project. All rights reserved.
+# Copyright (c) Contributors to the aswf-docker Project. All rights reserved.
+# SPDX-License-Identifier: MIT
+#
+# From: https://github.com/conan-io/conan-center-index/blob/5f19361f49bc31a349b383c1f95c7f79627f728a/recipes/nlohmann_json/all/conanfile.py
+
+from conan import ConanFile
+from conan.tools.build import check_min_cppstd
+from conan.tools.files import copy, get
+from conan.tools.layout import basic_layout
+import os
+
+required_conan_version = ">=1.50.0"
+
+
+class NlohmannJsonConan(ConanFile):
+    name = "nlohmann_json"
+    homepage = "https://github.com/nlohmann/json"
+    description = "JSON for Modern C++ parser and generator."
+    topics = "json", "header-only"
+    url = "https://github.com/conan-io/conan-center-index"
+    license = "MIT"
+    package_type = "header-library"
+    settings = "os", "arch", "compiler", "build_type"
+    no_copy_source = True
+
+    @property
+    def _minimum_cpp_standard(self):
+        return 11
+
+    def layout(self):
+        basic_layout(self, src_folder="src")
+
+    def package_id(self):
+        self.info.clear()
+
+    def validate(self):
+        if self.settings.compiler.cppstd:
+            check_min_cppstd(self, self._minimum_cpp_standard)
+
+    def source(self):
+        get(self, **self.conan_data["sources"][self.version], strip_root=True)
+
+    def generate(self):
+        pass
+
+    def build(self):
+        pass
+
+    def package(self):
+        # ASWF: license file in package specific directories
+        copy(self, "LICENSE*", self.source_folder, os.path.join(self.package_folder, "licenses", self.name))
+        copy(self, "*", os.path.join(self.source_folder, "include"), os.path.join(self.package_folder, "include"))
+
+    def package_info(self):
+        self.cpp_info.set_property("cmake_file_name", "nlohmann_json")
+        self.cpp_info.set_property("cmake_target_name", "nlohmann_json::nlohmann_json")
+        self.cpp_info.set_property("pkg_config_name", "nlohmann_json")
+        self.cpp_info.bindirs = []
+        self.cpp_info.libdirs = []